extract updateAndNotifyParentRoomWithParentMessage to lib file
